♻️ Refactor: 집안일 스텝1 페이지 로직, 스타일 분리
